reorganiza diretórios
