+++ .gitignore
@@ -137,11 +137,13 @@ dist
 # End of https://www.toptal.com/developers/gitignore/api/node
 
 # generated files
-src/graphql/generated/*
+src/generated/*
 
 .env*
 .flaskenv*
 !.env.project
 !.env.vault
 
+# TODO files
 *.todo
+*.todo.md
