+++ src/github-webhook/service/projectsV2Item/handleConverted.test.ts
@@ -1,8 +1,8 @@
-import { converted } from "../../payloads/board-issue/converted.js";
-import * as projectsV2Item from "../../../../github-gql/command/projectsV2Item.js";
+import { converted } from "../../../../test/webhook-payloads/board-issue/converted.js";
+import * as projectsV2Item from "../../../github-gql/issue.js";
 import { handleProjectsV2ItemConvertedEvent } from "./projectsV2Item.js";
-import { getIssueByNodeId } from "../../../../github-gql/command/projectsV2Item.js";
-import { mockPrisma } from "../../../../../test/mockPrisma.js";
+import { getIssueByNodeId } from "../../../github-gql/issue.js";
+import { mockPrisma } from "../../../../test/mockPrisma.js";
 
 describe("handleProjectsV2ItemConverted", () => {
   const contentNodeId = converted.projects_v2_item.content_node_id;
