+++ src/github-webhook/service/projectsV2Item/projectsV2Item.ts
@@ -3,9 +3,9 @@ import {
   ProjectsV2ItemCreatedEvent,
   ProjectsV2ItemEditedEvent,
 } from "@octokit/webhooks-types";
-import { isProjectsV2Issue } from "../../../../model/webhook/projectsV2Item.js";
-import { getIssueByNodeId } from "../../../../github-gql/command/projectsV2Item.js";
-import { IssueOperations } from "../../../../../prisma/operations/IssueOperations.js";
+import { isProjectsV2Issue } from "../../model/projectsV2Item.js";
+import { getIssueByNodeId } from "../../../github-gql/issue.js";
+import { IssueOperations } from "../../../../prisma/operations/IssueOperations.js";
 
 export async function handleProjectsV2ItemCreatedEvent({
   projects_v2_item,
