--- src/routes/github/index.ts
@@ -1,6 +1,0 @@
-import { Router } from "express";
-import { apolloRoutes } from "./apollo.js";
-
-export const githubRoutes = Router();
-
-githubRoutes.use("/apollo", apolloRoutes);
