--- src/routes/github/queries/BoardIssues.gql
@@ -1,69 +1,0 @@
-query BoardIssues {
-  organization(login: "wesleydecezere") {
-    id
-    name
-    projectV2(number: 27) {
-      title
-      number
-      creator {
-        login
-        url
-      }
-      items(first: 100) {
-        ... on ProjectV2ItemConnection {
-          issues: nodes {
-            content {
-              ... on Issue {
-                number
-                id
-                title
-                state
-                linkedBranches(first: 5) {
-                  nodes {
-                    id
-                  }
-                }
-              }
-            }
-            status: fieldValueByName(name: "Status") {
-              ... on ProjectV2ItemFieldSingleSelectValue {
-                name
-              }
-            }
-            tipo: fieldValueByName(name: "Tipo de Issue") {
-              ... on ProjectV2ItemFieldSingleSelectValue {
-                name
-              }
-            }
-            demanda: fieldValueByName(name: "Demanda") {
-              ... on ProjectV2ItemFieldSingleSelectValue {
-                name
-              }
-            }
-            repo: fieldValueByName(name: "Repository") {
-              ... on ProjectV2ItemFieldRepositoryValue {
-                repository {
-                  name
-                }
-              }
-            }
-            labels: fieldValueByName(name: "Labels") {
-              ... on ProjectV2ItemFieldLabelValue {
-                labels(first: 5) {
-                  nodes {
-                    name
-                  }
-                }
-              }
-            }
-            type
-          }
-        }
-        pageInfo {
-          endCursor
-          hasNextPage
-        }
-      }
-    }
-  }
-}
