--- src/routes/github/queries/LastIssues.gql
@@ -1,11 +1,0 @@
-query LastIssues($owner: String!, $repo: String!, $num: Int = 3) {
-  repository(owner: $owner, name: $repo) {
-    issues(last: $num) {
-      edges {
-        node {
-          title
-        }
-      }
-    }
-  }
-}
